Continue

+++ models/admins.model.js
@@ -19,26 +19,30 @@ async function adminLogin(email, password) {
                         data: {
                             _id: admin._id,
                         },
+                        status: 200
                     };
                 return {
                     msg: "Sorry, The Email Or Password Is Not Valid !!",
                     error: true,
                     data: {},
+                    status: 200
                 }
             }
             return {
-                msg: `Sorry, This Account Has Been Blocked !!`,
+                msg: "Sorry, This Account Has Been Blocked !!",
                 error: true,
                 data: {
                     blockingDate: admin.blockingDate,
                     blockingReason: admin.blockingReason,
                 },
+                status: 200
             }
         }
         return {
             msg: "Sorry, The Email Or Password Is Not Valid !!",
             error: true,
             data: {},
+            status: 200
         }
     }
     catch (err) {
@@ -48,19 +52,20 @@ async function adminLogin(email, password) {
 
 async function getAdminUserInfo(userId) {
     try {
-        // Check If User Is Exist
         const user = await adminModel.findById(userId);
         if (user) {
             return {
-                msg: `Get Admin Info For Id: ${user._id} Process Has Been Successfully !!`,
+                msg: "Get Admin Info Process Has Been Successfully !!",
                 error: false,
                 data: user,
+                status: 200
             }
         }
         return {
-            msg: "Sorry, The User Is Not Exist, Please Enter Another User Id !!",
+            msg: "Sorry, The User Is Not Exist !!",
             error: true,
             data: {},
+            status: 200
         }
     } catch (err) {
         throw Error(err);
@@ -72,23 +77,37 @@ async function getAdminsCount(merchantId, filters) {
         const merchant = await adminModel.findById(merchantId);
         if (merchant) {
             if (merchant.isMerchant) {
-                filters.storeId = merchant.storeId;
+                if (!merchant.isBlocked) {
+                    filters.storeId = merchant.storeId;
+                    return {
+                        msg: "Get Admins Count Process Has Been Successfully !!",
+                        error: false,
+                        data: await adminModel.countDocuments(filters),
+                        status: 200
+                    }
+                }
                 return {
-                    msg: "Get Admins Count Process Has Been Successfully !!",
-                    error: false,
-                    data: await adminModel.countDocuments(filters),
+                    msg: "Sorry, This Merchant Has Been Blocked !!",
+                    error: true,
+                    data: {
+                        blockingDate: merchant.blockingDate,
+                        blockingReason: merchant.blockingReason,
+                    },
+                    status: 401
                 }
             }
             return {
-                msg: "Sorry, Permission Denied !!",
+                msg: "Sorry, Permission Denied Because This Admin Is Not Merchant !!",
                 error: true,
                 data: {},
+                status: 401
             }
         }
         return {
             msg: "Sorry, This Merchant Is Not Exist !!",
             error: true,
             data: {},
+            status: 401
         }
     } catch (err) {
         throw Error(err);
@@ -99,24 +118,38 @@ async function getAllAdminsInsideThePage(merchantId, pageNumber, pageSize, filte
     try {
         const merchant = await adminModel.findById(merchantId);
         if (merchant) {
-            if (merchant.isMerchant && !merchant.isBlocked) {
-                filters.storeId = merchant.storeId;
+            if (merchant.isMerchant) {
+                if (!merchant.isBlocked) {
+                    filters.storeId = merchant.storeId;
+                    return {
+                        msg: `Get All Admins Inside The Page: ${pageNumber} Process Has Been Successfully !!`,
+                        error: false,
+                        data: await adminModel.find(filters).skip((pageNumber - 1) * pageSize).limit(pageSize).sort({ creatingDate: -1 }),
+                        status: 200
+                    }
+                }
                 return {
-                    msg: `Get All Admins Inside The Page: ${pageNumber} Process Has Been Successfully !!`,
-                    error: false,
-                    data: await adminModel.find(filters).skip((pageNumber - 1) * pageSize).limit(pageSize).sort({ creatingDate: -1 }),
+                    msg: "Sorry, This Merchant Has Been Blocked !!",
+                    error: true,
+                    data: {
+                        blockingDate: merchant.blockingDate,
+                        blockingReason: merchant.blockingReason,
+                    },
+                    status: 401
                 }
             }
             return {
-                msg: "Sorry, Permission Denied !!",
+                msg: "Sorry, Permission Denied Because This Admin Is Not Merchant !!",
                 error: true,
                 data: {},
+                status: 401
             }
         }
         return {
             msg: "Sorry, This Merchant Is Not Exist !!",
             error: true,
             data: {},
+            status: 401
         }
     } catch (err) {
         throw Error(err);
@@ -125,10 +158,10 @@ async function getAllAdminsInsideThePage(merchantId, pageNumber, pageSize, filte
 
 async function addNewAdmin(merchantId, adminInfo) {
     try{
-        const admin = await adminModel.findById(merchantId);
-        if (admin) {
-            if (admin.isMerchant){
-                if (!admin.isBlocked) {
+        const merchant = await adminModel.findById(merchantId);
+        if (merchant) {
+            if (merchant.isMerchant){
+                if (!merchant.isBlocked) {
                     const adminDetails = await adminModel.findOne({ email: adminInfo.email });
                     if (!adminDetails) {
                         (new adminModel({
@@ -212,33 +245,38 @@ async function addNewAdmin(merchantId, adminInfo) {
                             msg: "Create New Admin Process Has Been Successfully !!",
                             error: false,
                             data: {},
+                            status: 200
                         }
                     }
                     return {
                         msg: "Sorry, This Admin Is Already Exist !!",
                         error: true,
                         data: {},
+                        status: 200
                     }
                 }
                 return {
-                    msg: "Sorry, This Account Has Been Blocked !!",
+                    msg: "Sorry, This Merchant Has Been Blocked !!",
                     error: true,
                     data: {
-                        blockingDate: admin.blockingDate,
-                        blockingReason: admin.blockingReason,
+                        blockingDate: merchant.blockingDate,
+                        blockingReason: merchant.blockingReason,
                     },
+                    status: 401
                 }
             }
             return {
-                msg: "Sorry, Permission Denied !!",
+                msg: "Sorry, Permission Denied Because This Admin Is Not Merchant !!",
                 error: true,
                 data: {},
+                status: 401
             }
         }
         return {
-            msg: "Sorry, This Admin Is Not Exist !!",
+            msg: "Sorry, This Merchant Is Not Exist !!",
             error: true,
             data: {},
+            status: 401
         }
     }
     catch(err) {
@@ -248,43 +286,48 @@ async function addNewAdmin(merchantId, adminInfo) {
 
 async function updateAdminInfo(merchantId, adminId, newAdminDetails) {
     try {
-        const admin = await adminModel.findById(merchantId);
-        if (admin) {
-            if (admin.isMerchant){
-                if (!admin.isBlocked) {
+        const merchant = await adminModel.findById(merchantId);
+        if (merchant) {
+            if (merchant.isMerchant){
+                if (!merchant.isBlocked) {
                     const adminDetails = await adminModel.findOneAndUpdate({ _id: adminId }, newAdminDetails);
                     if (adminDetails) {
                         return {
                             msg: "Updating Admin Details Process Has Been Successfully !!",
                             error: false,
                             data: {},
+                            status: 200
                         }
                     }
                     return {
                         msg: "Sorry, This Admin Is Not Exist !!",
                         error: true,
                         data: {},
+                        status: 401
                     }
                 }
                 return {
-                    msg: "Sorry, This Account Has Been Blocked !!",
+                    msg: "Sorry, This Merchant Has Been Blocked !!",
                     error: true,
                     data: {
-                        blockingDate: admin.blockingDate,
-                        blockingReason: admin.blockingReason,
+                        blockingDate: merchant.blockingDate,
+                        blockingReason: merchant.blockingReason,
                     },
+                    status: 401
                 }
             }
             return {
-                msg: "Sorry, Permission Denied !!",
+                msg: "Sorry, Permission Denied Because This Admin Is Not Merchant !!",
                 error: true,
                 data: {},
+                status: 401
             }
         }
         return {
-            msg: "Sorry, This Admin Is Not Exist !!",
+            msg: "Sorry, This Merchant Is Not Exist !!",
             error: true,
             data: {},
+            status: 401
         }
     } catch (err) {
         throw Error(err);
@@ -296,37 +339,53 @@ async function deleteAdmin(merchantId, adminId){
         const admin = await adminModel.findById(merchantId);
         if (admin) {
             if (admin.isMerchant){
-                if ((new mongoose.Types.ObjectId(adminId)).equals(merchantId)) {
+                if (!admin.isBlocked) {
+                    if ((new mongoose.Types.ObjectId(adminId)).equals(merchantId)) {
+                        return {
+                            msg: "Sorry, Permission Denied Because This Admin Is Merchant !!",
+                            error: true,
+                            data: {},
+                            status: 401
+                        }
+                    }
+                    const adminDetails = await adminModel.findOneAndDelete({ _id: adminId });
+                    if (adminDetails) {
+                        return {
+                            msg: "Delete Admin Process Has Been Successfully !!",
+                            error: false,
+                            data: {},
+                            status: 200
+                        }
+                    }
                     return {
-                        msg: "Sorry, Permission Denied !!",
+                        msg: "Sorry, This Admin Is Not Exist !!",
                         error: true,
                         data: {},
-                    }
-                }
-                const adminDetails = await adminModel.findOneAndDelete({ _id: adminId });
-                if (adminDetails) {
-                    return {
-                        msg: "Delete Admin Process Has Been Successfully !!",
-                        error: false,
-                        data: {},
+                        status: 401
                     }
                 }
                 return {
-                    msg: "Sorry, This Admin Is Not Exist !!",
+                    msg: "Sorry, This Merchant Has Been Blocked !!",
                     error: true,
-                    data: {},
+                    data: {
+                        blockingDate: admin.blockingDate,
+                        blockingReason: admin.blockingReason,
+                    },
+                    status: 401
                 }
             }
             return {
-                msg: "Sorry, Permission Denied !!",
+                msg: "Sorry, Permission Denied Because This Admin Is Not Merchant !!",
                 error: true,
                 data: {},
+                status: 401
             }
         }
         return {
             msg: "Sorry, This Admin Is Not Exist !!",
             error: true,
             data: {},
+            status: 401
         }
     }
     catch(err){
